+++ docker/build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+docker build . -t orchestracities/grovepi:3.7-buster --platform linux/arm/v6