+++ RELEASE_NOTES.md
@@ -10,6 +10,9 @@
 - Ev Charging (#11)
 - Demo docker stack (#28)
 - Connected ttn sensors to lorawan agent (#16)
+- GrovePI base docker image
+- GrovePI example docker image
+- Connected sensors to json agent (#15)
 
 ### Bug fixes
 
@@ -18,7 +21,11 @@
 - Automated workflow for code / markdown linters
 - Automated workflow for Contributor License Agreement
 - Automated workflow for change log
+- Automated build of GrovePI base docker image
+- Automated build of GrovePI example docker image
 
 ### Documentation
+
+- GrovePI + Docker setup
 
 ### Technical debt